std: refactor storage
