+++ std/object/storage/dir_memory_fifo.go
@@ -1,4 +1,4 @@
-package object
+package storage
 
 import (
 	"sync"
