+++ std/object/storage/store_badger_test.go
@@ -1,12 +1,12 @@
 //go:build !js
 
-package object_test
+package storage_test
 
 import (
 	"os"
 	"testing"
 
-	"github.com/named-data/ndnd/std/object"
+	"github.com/named-data/ndnd/std/object/storage"
 	tu "github.com/named-data/ndnd/std/utils/testutils"
 	"github.com/stretchr/testify/require"
 )
@@ -17,7 +17,7 @@ func TestBadgerStore(t *testing.T) {
 	os.RemoveAll(dir)
 	defer os.RemoveAll(dir)
 
-	store, err := object.NewBadgerStore(dir)
+	store, err := storage.NewBadgerStore(dir)
 	require.NoError(t, err)
 	testStoreBasic(t, store)
 	testStoreRemoveRange(t, store)
