+++ std/object/storage/store_js_test.go
@@ -1,18 +1,18 @@
 //go:build js && wasm
 
-package object_test
+package storage_test
 
 import (
 	"syscall/js"
 	"testing"
 
-	"github.com/named-data/ndnd/std/object"
+	"github.com/named-data/ndnd/std/object/storage"
 	tu "github.com/named-data/ndnd/std/utils/testutils"
 )
 
 func TestJsStore(t *testing.T) {
 	tu.SetT(t)
-	store := object.NewJsStore(js.Global().Get("_ndnd_store_js"))
+	store := storage.NewJsStore(js.Global().Get("_ndnd_store_js"))
 	testStoreBasic(t, store)
 	testStoreRemoveRange(t, store)
 	// testStoreTxn(t, store)
