+++ std/object/storage/store_memory.go
@@ -1,4 +1,4 @@
-package object
+package storage
 
 import (
 	"fmt"
