+++ tools/catchunks.go
@@ -10,6 +10,7 @@ import (
 	"github.com/named-data/ndnd/std/log"
 	"github.com/named-data/ndnd/std/ndn"
 	"github.com/named-data/ndnd/std/object"
+	"github.com/named-data/ndnd/std/object/storage"
 	"github.com/spf13/cobra"
 )
 
@@ -51,7 +52,7 @@ func (cc *CatChunks) run(_ *cobra.Command, args []string) {
 	defer app.Stop()
 
 	// start object client
-	cli := object.NewClient(app, object.NewMemoryStore(), nil)
+	cli := object.NewClient(app, storage.NewMemoryStore(), nil)
 	err = cli.Start()
 	if err != nil {
 		log.Fatal(cc, "Unable to start object client", "err", err)
